+++ mobile/mysterium/openvpn_connection_setup.go
@@ -34,7 +34,6 @@ import (
 	"github.com/mysteriumnetwork/node/identity"
 	"github.com/mysteriumnetwork/node/nat/traversal"
 	"github.com/mysteriumnetwork/node/services/openvpn"
-	"github.com/mysteriumnetwork/node/services/openvpn/session"
 )
 
 const natPunchingMaxTTL = 10
@@ -76,7 +75,7 @@ func NewOpenVPNConnection(sessionTracker *sessionTracker, signerFactory identity
 
 		signer := signerFactory(options.ConsumerID)
 
-		username, password, err := session.SignatureCredentialsProvider(options.SessionID, signer)()
+		username, password, err := openvpn.SignatureCredentialsProvider(options.SessionID, signer)()
 		if err != nil {
 			return nil, nil, err
 		}
