+++ services/openvpn/service/auth_mocks.go
@@ -15,7 +15,7 @@
  * along with this program.  If not, see <http://www.gnu.org/licenses/>.
  */
 
-package session
+package service
 
 import (
 	"github.com/mysteriumnetwork/node/identity"
