+++ services/openvpn/service/server_config.go
@@ -15,7 +15,7 @@
  * along with this program.  If not, see <http://www.gnu.org/licenses/>.
  */
 
-package openvpn
+package service
 
 import (
 	"github.com/mysteriumnetwork/go-openvpn/openvpn/config"
